+++ WhoteachswhatJPA/src/main/java/com/seneca/repository/FacultyDao.java
@@ -8,15 +8,15 @@ public interface FacultyDao
 {
    public List<Faculty> getAll();
 
-   public Faculty getById(Long id);
+   public Faculty getById(Integer id);
 
    public void update(Faculty entity);
 
    public void create(Faculty entity);
 
-   public void delete(Long id);
+   public void delete(Integer id);
 
-   public List<Faculty> search(Faculty search, Long first, Integer maxItems);
+   public List<Faculty> search(Faculty search, Integer first, Integer maxItems);
 
    public Long getCount(Faculty search);
 }
