+++ WhoteachswhatJPA/src/main/java/com/seneca/repository/TeachingTypeDao.java
@@ -8,15 +8,15 @@ public interface TeachingTypeDao
 {
    public List<TeachingType> getAll();
 
-   public TeachingType getById(Long id);
+   public TeachingType getById(Integer id);
 
    public void update(TeachingType entity);
 
    public void create(TeachingType entity);
 
-   public void delete(Long id);
+   public void delete(Integer id);
 
-   public List<TeachingType> search(TeachingType search, Long first, Integer maxItems);
+   public List<TeachingType> search(TeachingType search, Integer first, Integer maxItems);
 
    public Long getCount(TeachingType search);
 }
